refactor: move TimedMetadata logic out of vidispine package
